Input Validation and Floating Labels
Re-merged the input validation code.
Form will now need valid input for the user to use the submit button.
Submit button will now redirect to index.php where it is expected to handle the form.
Introducting floating labels. Implemented for a more uniform look, especially with a lack in labels on the login page. Fully functioning!!!